Refactor reservation tests for get_reservations
